JWT 토큰 테스트

+++ demo/src/main/java/com/backend/demo/repository/DealRepository.java
@@ -4,5 +4,5 @@ import com.backend.demo.entity.Deal;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.repository.CrudRepository;
 
-public interface DealRepository extends JpaRepository<Deal, Integer> {
+public interface DealRepository extends JpaRepository<Deal, Long> {
 }
